requirements.txt
anthropic>=0.18.0
google-cloud-texttospeech>=2.14.0
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0
python-dotenv>=1.0.0
flask>=3.0.0
